+++ bindgen/ir/item.rs
@@ -1026,6 +1026,7 @@ impl Item {
                 FunctionKind::Method(MethodKind::Virtual { .. }) => {
                     cc.methods()
                 }
+                FunctionKind::Macro => true,
             },
         }
     }
@@ -1293,6 +1294,7 @@ impl Item {
             TypeKind::Void |
             TypeKind::Int(..) |
             TypeKind::Pointer(..) |
+            TypeKind::Function(..) |
             TypeKind::Float(..) => {}
             _ => panic!("Unsupported builtin type"),
         }